Create README.md
In this work, the image is divided into block and then apply BTC truncation coding. After that apply the NEQR approach to represent the image in the quantum domain. Then , measure the bpp and PSNR value to demonestrate the capability of the proposed QBTC approach compared to the NEQR approach.